requirements.txt
Django==1.11
Werkzeug==0.10.4
django-extensions==1.5.3
django-registration-redux==1.1
django-flat-theme==0.9.3
djangorestframework==3.6.4
djangorestframework-xml
django-bootstrap3==5.4.0
blist==1.3.6
cassandra-driver==3.6.0
django-gravatar2==1.4.2
-e git://github.com/livenson/libcdmi-python.git#egg=libcdmi.egg
requests==2.7
poster==0.8.1
alabaster==0.7.4
coverage==3.7.1
MarkupSafe==0.23
nose==1.3.6
Pygments==2.0.2
pytz==2015.2
PyYAML==3.11
shortuuid==0.4.2
gunicorn==19.3.0
crcmod==1.7
paho-mqtt==1.1
python-ldap==2.4.22
gremlinpython
djangodav
